requirements.txt
fastapi==0.115.5
langchain==0.3.9
langchain_community==0.3.8
mysql_connector_repackaged==0.3.1
numpy==2.1.3
openai==1.55.2
pandas==2.2.3
pydantic==2.10.2
python-dotenv==1.0.1
qdrant_client==1.12.1
Requests==2.32.3
streamlit==1.40.1
streamlit_option_menu==0.4.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
scikit-learn
matplotlib
